Large simplification of Jiraffet.
Most of the logic has been made to be synchronous. Each call has a reply
returned. This code is currently broken in that the run() method must be
changed to handle the timer functions of a node.

When a follower times out it must seek to be come a leader.

When a leader times out it must heartbeat.

Election timers are built into that synchronous call.

Time when a node is blocking for IO will be considered as times when the
node is partitioned from the cluster and the Raft algorithm should be
resilient enough to handle such interruptions.